OPAL-2082: Moved VariableEntityValueSetsDtoFunction to web-model, removing /_search from uriInfo path
