+++ opal-web-model/src/main/java/org/obiba/opal/web/magma/VariableEntityValueSetDtoFunction.java
@@ -1,4 +1,14 @@
-package org.obiba.opal.web.search.support;
+/*
+ * Copyright (c) 2013 OBiBa. All rights reserved.
+ *
+ * This program and the accompanying materials
+ * are made available under the terms of the GNU Public License v3.0.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.obiba.opal.web.magma;
 
 import org.obiba.magma.Value;
 import org.obiba.magma.ValueSet;
